server/requirements.txt
fastapi[standard]
tortoise-orm[asyncpg]
pydantic
python-dotenv
aerich
pwdlib[argon2]
PyJWT
pytest
pytest-asyncio
httpx
asgi_lifespan